RavenDB-19876 more refactoring
